requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
django==5.1.7
django-allauth==0.61.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-robots==6.1
django-taggit==5.0.1
executing==2.2.0
griffe==1.7.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==8.22.2
jedi==0.19.2
jiter==0.9.0
markdown==3.5.2
matplotlib-inline==0.1.7
mcp==1.6.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.69.0
openai-agents==0.0.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt-toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.1
pydantic-core==2.33.0
pydantic-settings==2.8.1
pygments==2.19.1
pyjwt==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-extensions==4.13.0
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13

# Exportação e Backup
celery==5.3.6
redis==5.0.1
openpyxl==3.1.2
reportlab==4.0.9
python-dateutil==2.8.2

# Otimização
django-redis==5.4.0
django-compressor==4.4
django-storages==1.14.2
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
sentry-sdk==1.39.1
